pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "parallax-maker"
version = "1.0.2"
description = "A workflow for turning images into 2.5D animations with depth-based segmentation and inpainting"
authors = [
    {name = "Niels Provos", email = "niels@provos.org"}
]
readme = "README.md"
license = "AGPL-3.0-or-later"
keywords = ["computer-vision", "depth-estimation", "image-segmentation", "3d-animation", "gltf", "dash", "stable-diffusion"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers", 
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: 3D Modeling",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "dash==2.17.0",
    "dash_extensions==1.0.14",
    "numpy==1.26.4",
    "Pillow>=10.3.0",
    "pygltflib==1.16.2",
    "torch>=2.2.2",
    "torchvision>=0.17.2",
    "opencv-python>=4.9.0",
    "timm==0.6.12",
    "diffusers[torch]>=0.29.0",
    "transformers>=4.40",
    "orjson>=3.10.1",
    "numba>=0.59.1",
    "websocket-client>=1.8.0",
    "scipy>=1.13.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
    "pre-commit",
]

[project.urls]
Homepage = "https://github.com/niels-provos/parallax-maker"
Documentation = "https://github.com/niels-provos/parallax-maker#readme"
Repository = "https://github.com/niels-provos/parallax-maker"
Issues = "https://github.com/niels-provos/parallax-maker/issues"

[project.scripts]
parallax-maker = "parallax_maker.webui:main"
parallax-gltf-cli = "parallax_maker.gltf_cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["parallax_maker"]
exclude = ["parallax_maker.assets*"]

[tool.setuptools.package-data]
parallax_maker = [
    "assets/**/*",
    "example/**/*",
    "*.js",
    "*.css", 
    "*.json",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=parallax_maker --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
